Cleanup portfolio components
