+++ apps/hyperdrive-trading/src/ui/portfolio/longs/usePortfolioLongsData.ts
@@ -9,7 +9,7 @@ import { getDrift } from "src/drift/getDrift";
 import { useAppConfigForConnectedChain } from "src/ui/appconfig/useAppConfigForConnectedChain";
 import { useAccount } from "wagmi";
 
-type OpenLongPositionsData = {
+export type OpenLongPositionsData = {
   hyperdrive: HyperdriveConfig;
   openLongs: OpenLongPositionReceived[];
 }[];
